requirements.txt
Package           Version
----------------- -------
appnope           0.1.2
asttokens         2.0.5
backcall          0.2.0
debugpy           1.6.0
decorator         5.1.1
entrypoints       0.4
executing         0.8.3
ipykernel         6.10.0
ipython           8.2.0
jedi              0.18.1
jupyter-client    7.2.0
jupyter-core      4.9.2
matplotlib-inline 0.1.3
nest-asyncio      1.5.4
numpy             1.22.3
pandas            1.4.1
parso             0.8.3
pexpect           4.8.0
pickleshare       0.7.5
pip               21.1.1
prompt-toolkit    3.0.28
psutil            5.9.0
ptyprocess        0.7.0
pure-eval         0.2.2
Pygments          2.11.2
python-dateutil   2.8.2
python-dotenv     0.20.0
pytz              2022.1
pyzmq             22.3.0
setuptools        56.0.0
six               1.16.0
stack-data        0.2.0
tornado           6.1
traitlets         5.1.1
typedframe        0.6.2
wcwidth           0.2.5
